Move editor to separate class
